fix(core): resolve all clippy warnings across frankenterm-core
Achieve zero warnings with `cargo clippy -p frankenterm-core --all-targets -- -D warnings`.
Fixes 178+ warnings across 145 files including src/, tests/, and benches/.

Warning categories resolved:
- float_cmp: replace assert_eq! on floats with epsilon-based comparisons
- unused must_use: prefix emit_trigger() calls with `let _ =`
- comparison_chain: rewrite if/else-if chains as match with Ordering
- approx_constant: use distinct non-PI values in test data
- clone_double_ref: dereference instead of cloning
- needless_collect: use iterators directly (.count(), .is_empty())
- needless_range_loop: replace index loops with enumerate()
- same_item_push: use extend(repeat_n(...))
- wildcard_matches_single_variant: name explicit variants
- similar_names: rename ambiguous bindings
- Auto-fixed via cargo clippy --fix: unused imports, redundant closures,
  manual_map, needless_borrow, and more

+++ crates/frankenterm-core/src/tantivy_reindex.rs
@@ -415,7 +415,7 @@ impl<W: IndexWriter> ReindexPipeline<W> {
                 }
 
                 // Range filter
-                if !range.includes(ordinal, record.event.occurred_at_ms as u64) {
+                if !range.includes(ordinal, record.event.occurred_at_ms) {
                     progress.events_filtered += 1;
                     last_offset = Some(record.offset.clone());
                     continue;
